fix(study-queue-rpc): cast stability/difficulty to NUMERIC to match RETURNS TABLE
PostgreSQL error 42804: fsrs_states.stability and .difficulty are
type `real` (float4) but the RETURNS TABLE declared them as NUMERIC.
PostgreSQL does not implicitly cast real → numeric in function returns.

Fix: explicit ::NUMERIC cast on COALESCE(cs.fsrs_stability, 1) and
COALESCE(cs.fsrs_difficulty, 5) in the final SELECT.

+++ supabase/migrations/20260303_03_study_queue_rpc.sql
@@ -175,8 +175,8 @@ BEGIN
     ROUND(cs.bkt_p_know::NUMERIC, 3) AS p_know,
     COALESCE(cs.fsrs_state_val, 'new') AS fsrs_state,
     cs.fsrs_due_at AS due_at,
-    COALESCE(cs.fsrs_stability, 1) AS stability,
-    COALESCE(cs.fsrs_difficulty, 5) AS difficulty,
+    COALESCE(cs.fsrs_stability, 1)::NUMERIC AS stability,
+    COALESCE(cs.fsrs_difficulty, 5)::NUMERIC AS difficulty,
     cs.card_is_new AS is_new,
     COUNT(*) OVER() AS total_count
   FROM cards_with_scores cs
